combine modules into common module
