+++ common/build.gradle.kts
@@ -1,0 +1,18 @@
+plugins {
+    id("org.example.age.java-conventions")
+    `java-library`
+}
+
+dependencies {
+    annotationProcessor("com.google.dagger:dagger-compiler:2.47")
+
+    implementation("com.google.dagger:dagger:2.47")
+    implementation("com.squareup.okhttp3:okhttp:4.11.0")
+    implementation("io.undertow:undertow-core:2.3.8.Final")
+    implementation("javax.inject:javax.inject:1")
+
+    testAnnotationProcessor("com.google.dagger:dagger-compiler:2.47")
+
+    testImplementation(project(":testing"))
+    testImplementation("com.squareup.okhttp3:mockwebserver:4.11.0")
+}
