+++ common/src/main/java/org/example/age/common/client/RequestModule.java
@@ -1,0 +1,12 @@
+package org.example.age.common.client;
+
+import dagger.Binds;
+import dagger.Module;
+
+/** Publishes a binding for {@link RequestDispatcher}. */
+@Module(includes = ClientModule.class)
+public interface RequestModule {
+
+    @Binds
+    RequestDispatcher bindRequestDispatcher(RequestDispatcherImpl impl);
+}
